+++ app/src/main/java/com/aiwazian/messenger/domain/DeleteMessagePayload.kt
@@ -9,5 +9,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DeleteMessagePayload(
     val chatId: Long,
-    val messageId: Int
+    val messageId: Long
 )
